+++ app/src/routes/events/[eventId]/conversations/index.jsx
@@ -212,6 +212,7 @@ export const EventConversationsPage = () => {
             thread={thread}
             messages={messages}
             responseRecipient={responseRecipient}
+            participants={participants}
             threadLoading={threadLoading}
             refetchThread={refetchThread}
             refetchThreads={refetchThreads}
